# src/RATools.Api/openapi.v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RATools.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/applications": {
      "get": {
        "tags": [
          "Applications"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Applications"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationRequestBody"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationRequestBody"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/applications/{id}": {
      "get": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "deleteMode",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ApplicationDeleteMode"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/applications/{id}/publish-history": {
      "get": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sequenceNumber",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "readinessStatus",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdFromUtc",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "createdToUtc",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationPublishHistoryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationPublishHistoryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationPublishHistoryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/applications/import": {
      "post": {
        "tags": [
          "Applications"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportApplicationRequestBody"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportApplicationRequestBody"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ImportApplicationRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationImportResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationImportResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationImportResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/applications/{id}/sequences": {
      "post": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSequenceRequestBody"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSequenceRequestBody"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSequenceRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/applications/{id}/sequences/{sequenceNumber}/publishing-metadata": {
      "get": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sequenceNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SequencePublishingMetadataDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SequencePublishingMetadataDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SequencePublishingMetadataDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sequenceNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSequencePublishingMetadataRequestBody"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSequencePublishingMetadataRequestBody"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSequencePublishingMetadataRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SequencePublishingMetadataDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SequencePublishingMetadataDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SequencePublishingMetadataDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/applications/{id}/sequences/{sequenceNumber}": {
      "delete": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sequenceNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deleteMode",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ApplicationDeleteMode"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApplicationDeleteMode": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "ApplicationDto": {
        "required": [
          "applicationNumber",
          "createdUtc",
          "ectdTemplateDisplayName",
          "ectdTemplateKey",
          "id",
          "sequences",
          "sponsorName",
          "workingDirectoryPath"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "applicationNumber": {
            "type": "string"
          },
          "sponsorName": {
            "type": "string"
          },
          "workingDirectoryPath": {
            "type": "string"
          },
          "createdUtc": {
            "type": "string",
            "format": "date-time"
          },
          "ectdTemplateKey": {
            "type": "string"
          },
          "ectdTemplateDisplayName": {
            "type": "string"
          },
          "sequences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SequenceDto"
            }
          }
        },
        "additionalProperties": false
      },
...
      "ApplicationImportResultDto": {
        "required": [
          "applicationId",
          "applicationNumber",
          "failedSequenceCount",
          "importedDocumentCount",
          "importedPlacementCount",
          "importedSequenceCount",
          "issues",
          "skippedSequenceCount",
          "workingDirectoryPath"
        ],
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "applicationNumber": {
            "type": "string"
          },
          "workingDirectoryPath": {
            "type": "string"
          },
          "importedSequenceCount": {
            "type": "integer",
            "format": "int32"
          },
          "importedDocumentCount": {
            "type": "integer",
            "format": "int32"
          },
          "importedPlacementCount": {
            "type": "integer",
            "format": "int32"
          },
          "skippedSequenceCount": {
            "type": "integer",
            "format": "int32"
          },
          "failedSequenceCount": {
            "type": "integer",
            "format": "int32"
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApplicationImportIssueDto"
            }
          }
        },
        "additionalProperties": false
      },
      "ApplicationPublishHistoryDto": {
        "required": [
          "applicationId",
          "applicationNumber",
          "entries",
          "lifecycleSummary",
          "page",
          "pageSize",
          "readinessSummary",
          "sponsorName",
          "statusSummary",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "applicationNumber": {
            "type": "string"
          },
          "sponsorName": {
            "type": "string"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "statusSummary": {
            "$ref": "#/components/schemas/ApplicationPublishHistoryStatusSummaryDto"
          },
          "readinessSummary": {
            "$ref": "#/components/schemas/ApplicationPublishHistoryReadinessAggregateDto"
          },
          "lifecycleSummary": {
            "$ref": "#/components/schemas/ApplicationPublishHistoryLifecycleSummaryDto"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApplicationPublishHistoryEntryDto"
            }
          }
        },
        "additionalProperties": false
      },
...
      "ApplicationPublishHistoryLifecycleSummaryDto": {
        "required": [
          "ambiguousCount",
          "appendTargetNotFoundCount",
          "currentSequenceCount",
          "deleteTargetNotFoundCount",
          "matchedCount",
          "replaceTargetNotFoundCount"
        ],
        "type": "object",
        "properties": {
          "matchedCount": {
            "type": "integer",
            "format": "int32"
          },
          "replaceTargetNotFoundCount": {
            "type": "integer",
            "format": "int32"
          },
          "deleteTargetNotFoundCount": {
            "type": "integer",
            "format": "int32"
          },
          "appendTargetNotFoundCount": {
            "type": "integer",
            "format": "int32"
          },
          "ambiguousCount": {
            "type": "integer",
            "format": "int32"
          },
          "currentSequenceCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ApplicationPublishHistoryReadinessAggregateDto": {
        "required": [
          "blockedCount",
          "readyCount",
          "unknownCount"
        ],
        "type": "object",
        "properties": {
          "readyCount": {
            "type": "integer",
            "format": "int32"
          },
          "blockedCount": {
            "type": "integer",
            "format": "int32"
          },
          "unknownCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ApplicationPublishHistoryStatusSummaryDto": {
        "required": [
          "completedCount",
          "failedCount",
          "runningCount"
        ],
        "type": "object",
        "properties": {
          "completedCount": {
            "type": "integer",
            "format": "int32"
          },
          "failedCount": {
            "type": "integer",
            "format": "int32"
          },
          "runningCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "CreateApplicationRequestBody": {
        "required": [
          "applicationNumber",
          "ectdTemplateKey",
          "sponsorName",
          "workingDirectoryParentPath"
        ],
        "type": "object",
        "properties": {
          "applicationNumber": {
            "maxLength": 32,
            "minLength": 3,
            "type": "string"
          },
          "ectdTemplateKey": {
            "maxLength": 32,
            "minLength": 2,
            "type": "string"
          },
          "sponsorName": {
            "maxLength": 128,
            "minLength": 2,
            "type": "string"
          },
          "workingDirectoryParentPath": {
            "maxLength": 512,
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "CreateSequenceRequestBody": {
        "required": [
          "description",
          "sequenceNumber",
          "submissionType"
        ],
        "type": "object",
        "properties": {
          "sequenceNumber": {
            "minLength": 1,
            "pattern": "^\\d{4}$",
            "type": "string"
          },
          "submissionType": {
            "maxLength": 64,
            "minLength": 2,
            "type": "string"
          },
          "description": {
            "maxLength": 512,
            "minLength": 2,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ImportApplicationRequestBody": {
        "required": [
          "ectdTemplateKey",
          "sponsorName",
          "workingDirectoryPath"
        ],
        "type": "object",
        "properties": {
          "workingDirectoryPath": {
            "maxLength": 1024,
            "minLength": 1,
            "type": "string"
          },
          "ectdTemplateKey": {
            "maxLength": 32,
            "minLength": 2,
            "type": "string"
          },
          "sponsorName": {
            "maxLength": 128,
            "minLength": 2,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "SequenceDto": {
        "required": [
          "createdUtc",
          "description",
          "sequenceNumber",
          "submissionType",
          "workingDirectoryPath"
        ],
        "type": "object",
        "properties": {
          "sequenceNumber": {
            "type": "string"
          },
          "submissionType": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "workingDirectoryPath": {
            "type": "string"
          },
          "createdUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "SequencePublishingMetadataDto": {
        "required": [
          "applicantContactName",
          "applicantContactType",
          "applicantName",
          "applicationId",
          "applicationType",
          "email",
          "formType",
          "sequenceDescription",
          "sequenceNumber",
          "standardsProfile",
          "submissionSubtype",
          "submissionType",
          "telephone",
          "telephoneNumberType"
        ],
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "sequenceNumber": {
            "type": "string"
          },
          "standardsProfile": {
            "type": "string"
          },
          "applicationType": {
            "type": "string",
            "nullable": true
          },
          "submissionType": {
            "type": "string"
          },
          "submissionSubtype": {
            "type": "string",
            "nullable": true
          },
          "sequenceDescription": {
            "type": "string"
          },
          "applicantName": {
            "type": "string"
          },
          "formType": {
            "type": "string",
            "nullable": true
          },
          "applicantContactName": {
            "type": "string",
            "nullable": true
          },
          "applicantContactType": {
            "type": "string",
            "nullable": true
          },
          "telephone": {
            "type": "string",
            "nullable": true
          },
          "telephoneNumberType": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateSequencePublishingMetadataRequestBody": {
        "required": [
          "applicantName",
          "sequenceDescription",
          "submissionType"
        ],
        "type": "object",
        "properties": {
          "applicationType": {
            "type": "string",
            "nullable": true
          },
          "submissionType": {
            "type": "string"
          },
          "submissionSubtype": {
            "type": "string",
            "nullable": true
          },
          "sequenceDescription": {
            "type": "string"
          },
          "applicantName": {
            "type": "string"
          },
          "formType": {
            "type": "string",
            "nullable": true
          },
          "applicantContactName": {
            "type": "string",
            "nullable": true
          },
          "applicantContactType": {
            "type": "string",
            "nullable": true
          },
          "telephone": {
            "type": "string",
            "nullable": true
          },
          "telephoneNumberType": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}